pyproject.toml
[project]
name = "videoforge"
version = "0.1.0"
description = "VideoForge local video editing automation pipeline"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [{name = "VideoForge Contributors"}]

dependencies = [
  "numpy>=1.24",
  "PyYAML>=6.0",
]

[project.optional-dependencies]
whisper = ["faster-whisper>=1.0.3"]
clip = ["open-clip-torch>=2.24.0"]
faiss = ["faiss-cpu>=1.8.0"]
resolve = []

[project.scripts]
videoforge = "VideoForge.cli.main:main"

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-q"
testpaths = ["tests"]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"
